% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{autocorrelation}
\alias{autocorrelation}
\title{compute correlation between aptamers in dataframe A and dataframe B}
\usage{
autocorrelation(
  df_a,
  df_b,
  corr_type = "spearman",
  comp_name = "correlation",
  sample_aptamers = 1000
)
}
\arguments{
\item{df_a}{dataframe A}

\item{df_b}{dataframe B.}

\item{corr_type}{correlation method (see ?correlate).}

\item{comp_name}{column name for correlation values.}

\item{sample_aptamers}{autocorrelate for N aptamers (1000)}
}
\value{
correlation dataframe
}
\description{
compute correlation between aptamers in dataframe A and dataframe B
}
\examples{
autocorrelation(table_1, table_2)
}
