% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biplotly.R
\name{biplotly}
\alias{biplotly}
\title{PCA biplot interactive with plotly}
\usage{
biplotly(table, x = "PC1", y = "PC2", color = "ARM", pca_obj = NULL)
}
\arguments{
\item{table}{matrix of aptamer values and metadata}

\item{x}{x value}

\item{y}{y value}

\item{color}{color by feature}

\item{pca_obj}{PCAtools::pca object}
}
\value{
plotly scatterplot of components
}
\description{
PCA biplot interactive with plotly
}
\examples{
biplotly(t_wdf$df, x="PC1", y="PC2", color="ARM", pca_obj = t_wdf$pca_obj)
}
