% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_plate_dist.R, R/visualization.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{Visualize 96-wells plate and sample positions.}
\usage{
draw_heatmap(table, color_by, text_by = NULL, title = "heatmap", text_size = 4)

draw_heatmap(table, color_by, text_by = NULL, title = "heatmap", text_size = 4)
}
\arguments{
\item{table}{digested_adat table.}

\item{color_by}{color samples by column.}

\item{text_by}{annotate samples by column.}

\item{title}{ggplot title.}

\item{text_size}{text fond size.}

\item{nrow}{TODO.}

\item{ncol}{TODO.}

\item{debug}{verbose (bool).}
}
\value{
ggplot class object

ggplot class object
}
\description{
Visualize 96-wells plate and sample positions.

Visualize 96-wells plate and sample positions.
}
\examples{
draw_heatmap(digested_adat$complete, "PlateId")
draw_heatmap(digested_adat$complete, "PlateId")
}
