% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_lm.R
\name{get_qqplot}
\alias{get_qqplot}
\title{generate a quantile-quantile plot from limma fitted model.}
\usage{
get_qqplot(model, coeff)
}
\arguments{
\item{model}{a limma class model.}

\item{coeff}{vector of coeffient names or indexes.}
}
\value{
ggplot class object.
}
\description{
generate a quantile-quantile plot from limma fitted model.
}
\examples{
get_qqplot(fit2, c("PBOvsTRT"))
}
