% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hemolyzed_plot.R
\name{hemo_plot}
\alias{hemo_plot}
\title{plot hemolyzed samples vs normal.}
\usage{
hemo_plot(accepted, rm_visual, rm_hba, hemo_id = "seq.4915.64")
}
\arguments{
\item{accepted}{table of accepted samples}

\item{rm_visual}{table of filtered by visual}

\item{rm_hba}{table of filtered by aptamer value}

\item{hemo_id}{aptamer colname}
}
\value{
boxplot of aptamer distribution by class
}
\description{
plot hemolyzed samples vs normal.
}
