% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_dist}
\alias{plot_dist}
\title{generate distribution plot aptamers}
\usage{
plot_dist(
  data,
  x = NULL,
  y = NULL,
  target = NULL,
  annotation = NULL,
  group_ids = NULL,
  rows_by = NULL,
  cols_by = NULL,
  sample_id = NULL,
  kind = "grid",
  title = "",
  x_lab = x,
  y_lab = y,
  x_rot = 90,
  debug = FALSE,
  scales_grid = "fixed",
  spaghetti = NULL,
  colspaghetti = NULL,
  stats = FALSE
)
}
\arguments{
\item{data}{tibble class object as samplexfeatures.}

\item{x}{x-axis value (def. aptamers).}

\item{y}{y-axis value (def. rfu).}

\item{target}{EntrezGeneSymbol or Aptamer SeqId.}

\item{annotation}{Somascan Annotation Table.}

\item{group_ids}{TODO.}

\item{rows_by}{facet_by row by feature (def. aptamers).}

\item{cols_by}{facet_by col by feature (def. groups).}

\item{kind}{if split plot with facet_by (def. grid)}

\item{title}{main plot title.}

\item{x_lab}{x axis label.}

\item{y_lab}{y axis label.}

\item{x_rot}{rotation x axis label (def. 90).}

\item{debug}{verbose level (bool).}

\item{scales_grid}{Are scales shared across all facets (the default, "fixed"), or do they vary across rows ("free_x"), columns ("free_y"), or both rows and columns ("free")?}

\item{spaghetti}{group samples by feature (allow spaghetti plot).}

\item{colspaghetti}{color spaghetti line by feature (def. NULL).}

\item{stats}{add stat_summary (median) (bool, def. FALSE).}
}
\value{
gglot object
}
\description{
generate distribution plot aptamers
}
\examples{
plot_dist(data=table,
annotation=anno, target='CXCL10',
x='time_idx', rows_by="aptamers",
cols_by='groups', kind="grid", x_rot=90,
scales_grid = 'free_y', spaghetti='SUBJID',
stats=TRUE)
}
