% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_distline}
\alias{plot_distline}
\title{Lineplot of calibrators/norm and alignments.}
\usage{
plot_distline(
  table,
  target = "calibrators",
  log2 = FALSE,
  title = target,
  x_lab = NULL,
  aptamer_tag = "seq.",
  sample_f = 0.2,
  log = TRUE
)
}
\arguments{
\item{table}{digested_adat table}

\item{target}{input is calibrators or raw/normalized RFU table}

\item{log2}{do log2 transformation of RFU values}

\item{title}{main plot title}

\item{x_lab}{label for x axis}

\item{sample_f}{subset table to build dist_plot norm def. 0.2 (float between 0 and 1)}

\item{log}{write to log file (def.TRUE)}
}
\value{
ggplot object.
}
\description{
Lineplot of calibrators/norm and alignments.
}
\examples{
plot_distline(digested_adat$calibrators, target="calibrators", log2=TRUE, title="calibrators", x_lab="RFU")
}
