% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_distribution}
\alias{plot_distribution}
\title{create distribution plot for aptamers}
\usage{
plot_distribution(table, title = NULL, colors_fix = NULL, color_by = NULL)
}
\arguments{
\item{table}{a table with expression values.}

\item{title}{main plot title.}

\item{colors_fix}{named vector with color groups.}

\item{color_by}{color by column value.}
}
\value{
distribution plot
}
\description{
create distribution plot for aptamers
}
\examples{
plot_distribution(digested_adat$complete, "a dist plot", color_by='ARMCD')
}
