% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_distscatter}
\alias{plot_distscatter}
\title{Scatterplot of two input tables.}
\usage{
plot_distscatter(
  x = table_a,
  y = table_b,
  sample_size = 5,
  title = "scatterplot",
  aptamers_name = NULL
)
}
\arguments{
\item{x}{first input table with aptamer seq columns}

\item{y}{second input table with aptamer seq columns}

\item{sample_size}{fit line on n sample (default=5)}

\item{title}{main plot title}

\item{aptamers_name}{vector of aptamers col names (extract from table if NULL)}
}
\value{
ggplot object.
}
\description{
Scatterplot of two input tables.
}
\examples{
plot_distscatter(digested_adat$complete_log2, digested_adat$postproces)
}
