% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_lm.R
\name{pval_distribution}
\alias{pval_distribution}
\title{generate a bar plot of p-values.}
\usage{
pval_distribution(model, coeff)
}
\arguments{
\item{model}{a limma class model.}

\item{coeff}{vector of coeffient names or indexes.}
}
\value{
plot class object.
}
\description{
generate a bar plot of p-values.
}
\examples{
pval_distribution(fit2, c("PBOvsTRT"))
}
