% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{A random generated somascan dataset of
142 samples and 7596 aptamers.}
\format{
A data frame with 142 rows and 7629 cols
}
\source{
hand-made dataset
}
\usage{
test_data
}
\description{
A random generated somascan dataset of
142 samples and 7596 aptamers.
}
\keyword{datasets}
