% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{wide_to_long_soma}
\alias{wide_to_long_soma}
\title{reshape somascan table from wide to long}
\usage{
wide_to_long_soma(
  data,
  x = NULL,
  y = NULL,
  target = NULL,
  annotation = NULL,
  group_ids = NULL,
  rows_by = NULL,
  cols_by = NULL,
  sample_id = "SampleId",
  spaghetti = NULL,
  colspaghetti = NULL
)
}
\arguments{
\item{data}{tibble class object as sample X features.}

\item{x}{x-axis value (def. aptamers).}

\item{y}{y-axis value (def. rfu).}

\item{target}{EntrezGeneSymbol or Aptamer SeqId.}

\item{annotation}{Somascan Annotation Table.}

\item{group_ids}{fill color of basic ggplot (aes) (def. NULL).}

\item{rows_by}{facet_by row by feature (def. aptamers).}

\item{cols_by}{facet_by col by feature (def. groups).}

\item{spaghetti}{group samples by feature (allow spaghetti plot).}

\item{colspaghetti}{color spaghetti line by feature (def. NULL).}
}
\value{
object tidy table long format
}
\description{
reshape somascan table from wide to long
}
\examples{
wide_to_long_soma(data=table,
annotation=anno, target='CXCL10',
x='time_idx', rows_by="aptamers",
cols_by='groups', spaghetti='SUBJID')
}
