% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplotly.R
\name{multiplotly}
\alias{multiplotly}
\title{multiplot PCs interactive with plotly}
\usage{
multiplotly(
  table,
  color = "ARM",
  pca_obj = NULL,
  marker_size = 4,
  show_legend = TRUE
)
}
\arguments{
\item{table}{matrix of aptamer values and metadata}

\item{color}{color by feature}

\item{pca_obj}{PCAtools::pca object}

\item{marker_size}{marker size}

\item{show_legend}{Def. TRUE}
}
\value{
multiple plotly scatterplot of components
}
\description{
multiplot PCs interactive with plotly
}
\examples{
multiplotly(t_wdf$df, color="ARM", pca_obj = t_wdf$pca_obj)
}
